requirements.txt
django<5
django-jalali-date==1.1.3
setuptools==75.5.0